+++ app/[[...params]]/page.tsx
@@ -1,0 +1,7 @@
+'use client'
+
+import HomePage from "@/app/pagey"
+
+export default function Page() {
+    return <HomePage />
+}
