+++ app/app.tsx
@@ -1,0 +1,10 @@
+import HomePage from "./pagey"
+import { BrowserRouter, Routes, Route } from "react-router-dom";
+
+export default function App() {
+    return <BrowserRouter>
+        <Routes>
+            <Route path="/" element={<HomePage />} />
+        </Routes>
+    </BrowserRouter>
+}
