+++ package.json
@@ -27,6 +27,8 @@
     "react": "^19.0.0",
     "react-dom": "^19.0.0",
     "react-markdown": "^10.1.0",
+    "react-router": "^6.30.1",
+    "react-router-dom": "^6.30.1",
     "react-syntax-highlighter": "^15.6.1",
     "remark-gfm": "^4.0.1",
     "sonner": "^2.0.5",
